Split OrderService into interface and implementation class
